requirements.txt
# Core RAG dependencies
python-dotenv>=1.0.0
pdfplumber>=0.10.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
sentence-transformers>=2.7.0
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
numpy>=1.24.0

# LLM
langchain-openai>=0.1.0
